id:1724 Fix unit tests to work with new auth schema
