+++ helpers/getApplicationList.js
@@ -1,0 +1,15 @@
+const UnitOfWork = require('../db');
+
+const getApplicationList = async () => {
+    const uow = new UnitOfWork();
+    const apps = await uow.applicationsRepository.getAllApplications();
+
+    const result = apps.reduce((map, app) => {
+        map[app.secretKey] = app;
+        return map;
+    }, {});
+
+    return result;
+};
+
+module.exports = {getApplicationList};
