+++ test/unit/applications.spec.js
@@ -62,6 +62,9 @@ describe('Applications API', () => {
             const options = {
                 url: '/api/applications/123456/userSignup',
                 method: 'POST',
+                headers: {
+                    'application-token': 'test'
+                },
                 payload: {
                     firstName:"admin",
                     lastName:"admin"
@@ -136,6 +139,9 @@ describe('Applications API', () => {
             const options = {
                 url: '/api/applications/d08a1f76-7c4a-4dd9-a377-83ffffa752f4/userSignup',
                 method: 'POST',
+                headers: {
+                    'application-token': 'test'
+                },
                 payload: {
                     primaryEmailAddress: "andrewrobb@sevenhillstechnology.com",
                     firstName: "admin",
@@ -242,6 +248,9 @@ describe('Applications API', () => {
             const options = {
                 url: '/api/applications/d08a1f76-7c4a-4dd9-a377-83ffffa752f4/userSignup',
                 method: 'POST',
+                headers: {
+                    'application-token': 'test'
+                },
                 payload: {
                     primaryEmailAddress: "andrewrobb@sevenhillstechnology.com",
                     firstName: "admin",
@@ -344,6 +353,9 @@ describe('Applications API', () => {
             const options = {
                 url: '/api/applications/d08a1f76-7c4a-4dd9-a377-83ffffa752f4/userSignup',
                 method: 'POST',
+                headers: {
+                    'application-token': 'test'
+                },
                 payload: {
                     primaryEmailAddress: "andrewrobb@sevenhillstechnology.com",
                     firstName: "admin",
@@ -442,6 +454,9 @@ describe('Applications API', () => {
             const options = {
                 url: '/api/applications/d08a1f76-7c4a-4dd9-a377-83ffffa752f4/userSignup',
                 method: 'POST',
+                headers: {
+                    'application-token': 'test'
+                },
                 payload: {
                     primaryEmailAddress: "andrewrobb@sevenhillstechnology.com",
                     firstName: "admin",
@@ -549,6 +564,9 @@ describe('Applications API', () => {
             const options = {
                 url: '/api/applications/d08a1f76-7c4a-4dd9-a377-83ffffa752f4/userSignup',
                 method: 'POST',
+                headers: {
+                    'application-token': 'test'
+                },
                 payload: {
                     primaryEmailAddress: "andrewrobb@sevenhillstechnology.com",
                     firstName: "admin",
